requirements.txt
asgiref==3.9.1
boto3==1.39.14
botocore==1.39.14
Django==4.2.23
django-allauth==65.9.0
django-browser-reload==1.18.0
django-cleanup==9.0.0
django-colorfield==0.14.0
django-environ==0.12.0
django-htmx==1.23.2
django-storages==1.14.6
django-tenants==3.8.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==1.26.20
whitenoise==6.9.0
